Standardize code-formatting on Goethe
