+++ conjure-java-server-verifier/src/test/java/com/palantir/conjure/java/verification/server/undertest/JerseyServerUnderTestApplication.java
@@ -41,6 +41,7 @@ public final class JerseyServerUnderTestApplication extends Application<JerseySe
     }
 
     @Override
+    @SuppressWarnings("ProxyNonConstantType")
     public void run(JerseyServerUnderTestConfiguration _configuration, Environment environment) {
         environment
                 .jersey()
