+++ conjure-java-server-verifier/src/test/java/com/palantir/conjure/java/verification/server/undertest/UndertowServerUnderTestExtension.java
@@ -34,6 +34,7 @@ public final class UndertowServerUnderTestExtension implements BeforeAllCallback
     private Undertow server;
 
     @Override
+    @SuppressWarnings("ProxyNonConstantType")
     public void beforeAll(ExtensionContext _context) {
         UndertowAutoDeserializeService autoDeserialize =
                 Reflection.newProxy(UndertowAutoDeserializeService.class, new EchoResourceInvocationHandler());
