+++ versions.lock
@@ -3,22 +3,20 @@ com.atlassian.commonmark:commonmark:0.12.1 (1 constraints: 36052a3b)
 com.fasterxml.jackson:jackson-bom:2.12.3 (14 constraints: 34376253)
 com.fasterxml.jackson.core:jackson-annotations:2.12.3 (12 constraints: aed0cc01)
 com.fasterxml.jackson.core:jackson-core:2.12.3 (16 constraints: 154b8b84)
-com.fasterxml.jackson.core:jackson-databind:2.12.3 (29 constraints: ee2b7c08)
+com.fasterxml.jackson.core:jackson-databind:2.12.3 (28 constraints: e718723c)
 com.fasterxml.jackson.dataformat:jackson-dataformat-cbor:2.12.3 (3 constraints: 9742dc9f)
 com.fasterxml.jackson.dataformat:jackson-dataformat-smile:2.12.3 (2 constraints: 352b4987)
 com.fasterxml.jackson.dataformat:jackson-dataformat-yaml:2.12.3 (4 constraints: ec33db91)
 com.fasterxml.jackson.datatype:jackson-datatype-guava:2.12.3 (3 constraints: e33991fc)
-com.fasterxml.jackson.datatype:jackson-datatype-jdk8:2.12.3 (7 constraints: af6d2445)
+com.fasterxml.jackson.datatype:jackson-datatype-jdk8:2.12.3 (6 constraints: a85aa2c5)
 com.fasterxml.jackson.datatype:jackson-datatype-joda:2.12.3 (3 constraints: e33991fc)
 com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.12.3 (3 constraints: e33991fc)
 com.fasterxml.jackson.module:jackson-module-afterburner:2.12.3 (3 constraints: ff393303)
-com.fasterxml.jackson.module:jackson-module-parameter-names:2.12.3 (3 constraints: 6d305ba2)
 com.github.ben-manes.caffeine:caffeine:2.9.0 (9 constraints: aba0f467)
 com.google.code.findbugs:jsr305:3.0.2 (20 constraints: a13e998f)
 com.google.errorprone:error_prone_annotations:2.7.1 (17 constraints: 6f1135ed)
-com.google.errorprone:javac-shaded:9+181-r4173-1 (1 constraints: b1140fc6)
 com.google.guava:failureaccess:1.0.1 (1 constraints: 140ae1b4)
-com.google.guava:guava:30.1.1-jre (32 constraints: 584e189b)
+com.google.guava:guava:30.1.1-jre (30 constraints: eb236cb3)
 com.google.guava:listenablefuture:9999.0-empty-to-avoid-conflict-with-guava (1 constraints: bd17c918)
 com.google.j2objc:j2objc-annotations:1.3 (1 constraints: b809eda0)
 com.palantir.common:streams:1.9.1 (1 constraints: 0d050c36)
@@ -38,8 +36,7 @@ com.palantir.dialogue:dialogue-core:3.1.0 (3 constraints: ff3cec67)
 com.palantir.dialogue:dialogue-futures:3.1.0 (3 constraints: 23339033)
 com.palantir.dialogue:dialogue-serde:3.1.0 (3 constraints: 3d2e8252)
 com.palantir.dialogue:dialogue-target:3.1.0 (7 constraints: 26753bec)
-com.palantir.javaformat:palantir-java-format:1.1.0 (1 constraints: 0405f335)
-com.palantir.javaformat:palantir-java-format-spi:1.1.0 (1 constraints: d5120326)
+com.palantir.goethe:goethe:0.1.0 (1 constraints: 0305ee35)
 com.palantir.refreshable:refreshable:2.0.1 (2 constraints: 62243ab7)
 com.palantir.ri:resource-identifier:1.3.0 (3 constraints: b82443f1)
 com.palantir.safe-logging:preconditions:1.16.0 (23 constraints: f28ebf6a)
@@ -54,7 +51,7 @@ com.palantir.tritium:tritium-api:0.22.0 (2 constraints: 291ffebb)
 com.palantir.tritium:tritium-core:0.22.0 (1 constraints: 3c1036a2)
 com.palantir.tritium:tritium-metrics:0.22.0 (5 constraints: 565c7623)
 com.palantir.tritium:tritium-registry:0.22.0 (9 constraints: 8faad6bf)
-com.squareup:javapoet:1.13.0 (1 constraints: 3705323b)
+com.squareup:javapoet:1.13.0 (2 constraints: 2b113eee)
 commons-codec:commons-codec:1.15 (1 constraints: 0d13c328)
 info.picocli:picocli:4.6.1 (1 constraints: 0d051236)
 io.dropwizard.metrics:metrics-core:4.1.2 (19 constraints: ee2d9f64)
@@ -68,7 +65,6 @@ org.apache.httpcomponents.client5:httpclient5:5.1 (1 constraints: 6c132e47)
 org.apache.httpcomponents.core5:httpcore5:5.1.1 (3 constraints: 5a39f50b)
 org.apache.httpcomponents.core5:httpcore5-h2:5.1.1 (1 constraints: 3b13073c)
 org.checkerframework:checker-qual:3.10.0 (2 constraints: 461ab359)
-org.functionaljava:functionaljava:4.8 (1 constraints: 81129900)
 org.glassfish.hk2:osgi-resource-locator:1.0.3 (2 constraints: 7d23b265)
 org.glassfish.hk2.external:jakarta.inject:2.6.1 (9 constraints: fd9540f6)
 org.glassfish.jersey.core:jersey-common:2.34 (10 constraints: 25b1dcc3)
@@ -94,6 +90,7 @@ com.fasterxml.jackson.jaxrs:jackson-jaxrs-base:2.12.3 (3 constraints: 903d4249)
 com.fasterxml.jackson.jaxrs:jackson-jaxrs-cbor-provider:2.12.3 (2 constraints: f7276e28)
 com.fasterxml.jackson.jaxrs:jackson-jaxrs-json-provider:2.12.3 (2 constraints: 6e1d2e4b)
 com.fasterxml.jackson.module:jackson-module-jaxb-annotations:2.12.3 (3 constraints: 903d4249)
+com.fasterxml.jackson.module:jackson-module-parameter-names:2.12.3 (2 constraints: 661d2d2f)
 com.helger:profiler:1.1.1 (1 constraints: e21053b8)
 com.jcraft:jzlib:1.1.3 (1 constraints: 0d192ea2)
 com.netflix.concurrency-limits:concurrency-limits-core:0.2.2 (1 constraints: 7d14fe7b)
